Complete core documents and enable working RTD pipeline
This commit completes the initial milestone of a completely-working pipeline.
The repository contains all core files (except Conventions, which will be done
after several categories have been completed), the necessary Sphinx config
files and build settings, and a working integration with ReadTheDocs.
From this point on, the majority of the work is just filling out the categories
from the latest version of the Topical TF PDF, and the intermediate content in
the POC MarkDown version of this repository.